+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 
-## Unreleased
+## 1.4.1 - 2016-12-16
 
 ### Fixed
 
@@ -15,6 +15,7 @@
 - Message, stream and URI factories for [Slim Framework](https://github.com/slimphp/Slim)
 - BufferedStream that allow you to decorate a non-seekable stream with a seekable one.
 - cUrlFormatter to be able to redo the request with a cURL command
+
 
 ## 1.3.1 - 2016-07-15